HADOOP-13786 S3A committers don't need to support a JobContext in the constructors or factories: remove, clean up tests. Where tests do need to create a Committer with nothing but a JobConf, use the same code which MR itself does for this, now statically exported from AbstractCommitITest

+++ hadoop-tools/hadoop-aws/src/test/java/org/apache/hadoop/fs/s3a/commit/staging/TestStagingCommitter.java
@@ -151,12 +151,13 @@ public class TestStagingCommitter extends StagingTestBase.MiniDFSTest {
     // and bind the FS
     wrapperFS.setAmazonS3Client(mockClient);
 
-    this.job = new JobContextImpl(jobConf, JOB_ID);
-    this.jobCommitter = new MockedStagingCommitter(OUTPUT_PATH, job);
-    jobCommitter.setupJob(job);
 
+    this.job = new JobContextImpl(jobConf, JOB_ID);
     this.tac = new TaskAttemptContextImpl(
         new Configuration(job.getConfiguration()), AID);
+
+    this.jobCommitter = new MockedStagingCommitter(OUTPUT_PATH, tac);
+    jobCommitter.setupJob(job);
 
     // get the task's configuration copy so modifications take effect
     this.conf = tac.getConfiguration();
@@ -177,7 +178,9 @@ public class TestStagingCommitter extends StagingTestBase.MiniDFSTest {
   @After
   public void cleanup() {
     try {
-      FileUtils.deleteDirectory(tmpDir);
+      if (tmpDir != null) {
+        FileUtils.deleteDirectory(tmpDir);
+      }
     } catch (IOException ignored) {
 
     }
